Bundle as CommonJS.
Also strips unnecessary dependencies, e.g. those supporting jest (which
has been replaced by mocha).
