--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [["@babel/preset-env", {"modules": false}]],
-  "env": {
-    "test": {
-      "presets": ["@babel/preset-env"]
-    }
-  }
-}
