+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "urbit-keygen",
-  "version": "2.1.0",
-  "description": "Wrapper functions for key generation according to the Urbit hierarchical wallet design.",
+  "version": "2.3.0",
+  "description": "Key derivation and HD wallet generation functions for Urbit.",
   "main": "src/index.js",
   "scripts": {
-    "test": "jest --coverage tests/test.js && mocha tests/property.js",
-    "build": "mkdir dist && browserify src/index.js -s urbit-keygen > dist/index.js"
+    "test": "nyc mocha --reporter spec"
   },
   "repository": {
     "type": "git",
@@ -25,19 +24,13 @@
     "bip32": "^1.0.2",
     "isomorphic-webcrypto": "^1.6.0",
     "lodash": "^4.17.11",
-    "urbit-ob": "^1.4.1",
+    "urbit-ob": "^2.0.0",
     "tweetnacl": "^1.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0-beta.34",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^23.4.2",
-    "browserify": "^16.2.2",
     "chai": "^4.2.0",
-    "jest": "^23.5.0",
     "jsverify": "^0.8.3",
     "mocha": "^5.2.0",
-    "regenerator-runtime": "^0.12.1"
+    "nyc": "^13.1.0"
   }
 }
